Move all constants to the same file
