--- assets/src/blocks/amp-story/constants.js
@@ -1,6 +1,0 @@
-
-export const IMAGE_BACKGROUND_TYPE = 'image';
-export const VIDEO_BACKGROUND_TYPE = 'video';
-export const ALLOWED_MEDIA_TYPES = [ 'image', 'video' ];
-export const POSTER_ALLOWED_MEDIA_TYPES = [ 'image' ];
-export const MEDIA_INNER_BLOCKS = [ 'core/video', 'core/audio' ];
