+++ assets/src/blocks/amp-story/index.js
@@ -8,9 +8,8 @@ import { InnerBlocks } from '@wordpress/editor';
 /**
  * Internal dependencies
  */
-import { BLOCK_ICONS } from '../../constants';
+import { BLOCK_ICONS, IMAGE_BACKGROUND_TYPE, VIDEO_BACKGROUND_TYPE } from '../../constants';
 import EditPage from './edit';
-import { IMAGE_BACKGROUND_TYPE, VIDEO_BACKGROUND_TYPE } from './constants';
 
 export const name = 'amp/amp-story-page';
